Add Abstract Entity with auditable fields and perform restructuring entities
