--- src/main/java/com/mangalaxy/mango/domain/Degree.java
@@ -1,11 +1,0 @@
-package com.mangalaxy.mango.domain;
-
-/**
- * Defines types of different educational degree levels.
- */
-public enum Degree {
-  ASSOCIATE,  // Associate degree
-  BACHELOR,   // Bachelor's degree
-  MASTER,     // Master's degree
-  PHD         // Doctoral degree
-}
