--- src/main/java/com/mangalaxy/mango/domain/LangLevel.java
@@ -1,11 +1,0 @@
-package com.mangalaxy.mango.domain;
-
-/**
- * Enumerates language proficiency levels.
- */
-public enum LangLevel {
-  ELEMENTARY,
-  INTERMEDIATE,
-  ADVANCED,
-  FLUENT
-}
