+++ src/main/java/com/mangalaxy/mango/domain/entity/Talent.java
@@ -1,14 +1,17 @@
 package com.mangalaxy.mango.domain.entity;
 
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
 import org.hibernate.annotations.NaturalId;
 import org.hibernate.annotations.NaturalIdCache;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
-import javax.persistence.ForeignKey;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
@@ -17,6 +20,9 @@ import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Size;
 
 @Data
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
 @EqualsAndHashCode(callSuper = true)
 @Entity
 @Table(name = "talent")
@@ -39,8 +45,7 @@ public class Talent extends AbstractEntity {
   private String password;
 
   @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "location_id", referencedColumnName = "id",
-        foreignKey = @ForeignKey(name = "location_id_fk"))
+  @JoinColumn(name = "location_id", referencedColumnName = "id")
   private Location location;
 
 }
