+++ src/main/resources/application.properties
@@ -1,8 +1,7 @@
 # Server port
 server.port=9000
-
-# DATASOURCE (DataSourceAutoConfiguration & DataSourceProperties)
-spring.datasource.url=jdbc:postgresql://localhost:5432/mangodb
+# Datasource (DataSourceAutoConfiguration & DataSourceProperties)
+spring.datasource.url=jdbc:postgresql://localhost:5432/mango
 spring.datasource.username=postgres
 spring.datasource.password=postgres
 
@@ -14,6 +13,7 @@ spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation=true
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 # only for development mode, change it for production stage
 spring.jpa.hibernate.ddl-auto=create
+# Logging levels
 logging.level.org.hibernate.SQL=DEBUG
 logging.level.org.hibernate.type=TRACE
